requirements.txt
huggingface_hub>=0.24.0
pytest>=8.0.0
coverage>=7.0
typing_extensions>=4.10
openai>=1.0.0
python-dotenv>=1.0.0
Flask>=3.0.0
Flask-CORS>=4.0.0
boto3>=1.28.0
mangum>=0.17.0
jsonschema>=4.17.0
requests>=2.31.0
asgiref>=3.7.0
